+++ Sources/Views/ALKMyFormCell.swift
@@ -73,6 +73,8 @@ class ALKMyFormCell: ALKFormCell {
                                                       padding: ViewPadding.messageViewPadding)
         if !isMessageEmpty {
             messageView.update(model: model)
+        } else if #available(iOS 17, *) {
+            messageView.update(model: model)
         }
 
         messageView.updateHeighOfView(hideView: isMessageEmpty, model: model)