+++ Sources/Views/KMFormDropDownCell.swift
@@ -11,6 +11,7 @@ import iOSDropDown
 
 protocol KMFormDropDownSelectionProtocol {
     func optionSelected(position: Int,selectedText: String?,index:Int)
+    func defaultOptionSelected(position: Int,selectedText: String?,index:Int)
 }
 
 class KMFormDropDownCell: UITableViewCell {
@@ -66,6 +67,7 @@ class KMFormDropDownCell: UITableViewCell {
             for (index, item) in options.enumerated() {
                 if let selected = item.selected, selected {
                     selectedIndex = index
+                    addDefultDataInCache(position: self.menu.tag, selectedText: item.value, index: index)
                 }
             }
             
@@ -89,6 +91,7 @@ class KMFormDropDownCell: UITableViewCell {
     var name: String = ""
     var optionsDict: [String: FormTemplate.Option] = [:]
     var delegate: KMFormDropDownSelectionProtocol?
+    var identifier: String?
     
     let nameLabel: UILabel = {
         let label = UILabel(frame: .zero)
@@ -134,6 +137,12 @@ class KMFormDropDownCell: UITableViewCell {
         fatalError("init(coder:) has not been implemented")
     }
 
+    private func addDefultDataInCache(position: Int,selectedText: String?,index:Int) {
+        DispatchQueue.main.async {
+            self.delegate?.defaultOptionSelected(position: position, selectedText: selectedText, index: index)
+        }
+    }
+    
     private func addConstraints() {
         addViewsForAutolayout(views: [view, nameLabel, errorStackView])
         errorStackView.addArrangedSubview(errorLabel)